Revert "Bgk 63 fix"

+++ BGK.Data/Entities/Company.cs
@@ -8,7 +8,7 @@ namespace BearGoodbyeKolkhozProject.Data.Entities
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public long Tin { get; set; }
+        public int Tin { get; set; }
         public string? Email { get; set; }
         public string? PhoneNumber { get; set; }
         public string Password { get; set; }
